ai-models/requirements.txt
# Core AI/ML Libraries
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
diffusers==0.24.0
accelerate==0.24.0

# Image Processing
Pillow==10.0.1
opencv-python==4.8.1.78
scikit-image==0.21.0
imageio==2.31.5

# Stable Diffusion
stable-diffusion-2==1.0.0
xformers==0.0.22.post7

# Style Transfer
neural-style-transfer==1.0.0
tensorflow==2.14.0

# Utilities
numpy==1.24.3
scipy==1.11.3
matplotlib==3.7.2
seaborn==0.12.2

# API Integration
requests==2.31.0
aiohttp==3.8.6
websockets==11.0.3

# Storage
boto3==1.34.0
minio==7.1.17

# Configuration
python-dotenv==1.0.0
pydantic==2.4.2

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1 